[cleanup] GAME uses prevProps vs state.currMove, moved all actions to actions/index.js
