--- ui/client/src/actions/action-castle-king.js
@@ -1,8 +1,0 @@
-const castleKing = (rowStart, colStart, rowEnd, colEnd, pieceToMove) => {
-  return {
-    type: 'CASTLING',
-    payload: {rowStart, colStart, rowEnd, colEnd, pieceToMove},
-  };
-}
-
-export default castleKing;
