--- ui/client/src/actions/action-init-game.js
@@ -1,8 +1,0 @@
-const initGame = (id, white, black, position, whiteToMove, moves) => {
-  return {
-    type: 'GAME_INITIALIZED',
-    payload: { id, white, black, position, whiteToMove, moves },
-  };
-}
-
-export default initGame;
