--- ui/client/src/actions/action-load-games.js
@@ -1,8 +1,0 @@
-const loadGames = (games) => {
-  return {
-    type: 'GAMES_ACCESSED',
-    payload: games
-  };
-}
-
-export default loadGames;
