--- ui/client/src/actions/action-store-user.js
@@ -1,8 +1,0 @@
-const storeUser = (userId) => {
-  return {
-    type: 'LOGGED_IN',
-    payload: userId
-  }
-}
-
-export default storeUser;
