--- ui/client/src/actions/action-toggle-turn.js
@@ -1,9 +1,0 @@
-const toggleTurn = () => {
-  console.log('\tturn toggled!!!')
-  return {
-    type: 'PLAYER_MOVED',
-    payload: null
-  };
-}
-
-export default toggleTurn;
