--- ui/client/src/actions/action-update-check-status.js
@@ -1,10 +1,0 @@
-const updateCheckStatus = (userId) => {
-  console.log('in action:', userId)
-  return {
-    type: 'CHECK_STATUS_UPDATED',
-    payload: userId,
-  }
-}
-
-export default updateCheckStatus
-
