--- ui/client/src/actions/action-update-position.js
@@ -1,8 +1,0 @@
-const updatePosition = (origin, destination, pieceToMove) => {
-  return {
-    type: 'POSITION_CHANGED',
-    payload:  [ origin, destination, pieceToMove ]  
-  };
-}
-
-export default updatePosition;
